GradCam Pipeline IGF Pipeline XRAI Pipeline
Made IGF and XRAI Pipeline.
Adapted GradCAM pipeline to accept any model.
The target layer can be given as a string now and is parsed depending on weather a layer is specified or an int is given.
e.g. for convnext which in the notebook was
GRADCAM_TARGET_LAYER = model.features[-1][-1].block[0] # convnext
you can give --target_layer "features.-1.-1.block.0"
as argument!