reverted changes from refactoring
reverted change in the transforms which changed how the XAI heatmaps were derived back from crops to a resize. In the future I would recommend using Resize and Centercrop again, but this requires to have consistency with the other transforms and the creation of the datasets.
Possibly it makes sense to resize + centercrop or centercrop the testset at creation to fix the slight missmatch in size of the street images in the current setup.